Re-write and Optimization
Merged this with other personal projects that I am working on to output results to files on my local machine. After that I stripped out that extra set of methods to allow this to run independently on Google Sheets. Building a deployment version as well as a stand-alone test sheet for others to use. Some areas still need some work, but it is getting better and is more optimized than previous versions.